feat: Phase 5 - Task 6.3 完了 (jira/confluence/workflow コマンド移行)
## 変更内容

### ✅ Task 6.3: 他のコマンドハンドラーの移行（完了）

#### 1. jira-sync コマンドの移行
- `src/presentation/commands/jira/sync.ts` (42行) - JIRA同期コマンド
- `scripts/jira-sync.ts` (30行) - 薄いラッパー

#### 2. confluence-sync コマンドの移行
- `src/presentation/commands/confluence/sync.ts` (36行) - Confluence同期コマンド
- `scripts/confluence-sync.ts` (37行) - 薄いラッパー

#### 3. workflow-orchestrator コマンドの移行
- `src/presentation/commands/workflow/orchestrator.ts` (235行) - ワークフローオーケストレーター
- `scripts/workflow-orchestrator.ts` (53行) - 薄いラッパー

## テスト結果

✅ **ビルド成功**
\`\`\`
npm run build
\`\`\`

✅ **全テスト成功**
\`\`\`
npm test
Test Suites: 925 passed, 5 skipped
\`\`\`

## 後方互換性

すべての既存のインポートパスは引き続き動作します。
